dataowner/requirements.txt
numpy==1.23.2
Flask==2.1.3
scikit-learn==1.1.2
pandas==1.4.3
requests==2.22.0
python-dotenv>=1.0.0
retry==0.9.2
coloredlogs==15.0.1
--extra-index-url https://test.pypi.org/simple/ 
mictlanx==0.0.57
rory==0.0.61